static_cast mass attack

+++ sfml template/Actor.cpp
@@ -40,7 +40,7 @@ void Actor::takeControl(sf::Event event, Map& map, sf::RenderWindow& w, std::vec
 	if (sf::Mouse::isButtonPressed(sf::Mouse::Right)) {
 		if (event.MouseButtonReleased) {
 			if (this->towns.size() > 0) {
-				for (int i = 0; i < towns.size(); i++) {
+				for (int i = 0; i < static_cast<int>(towns.size()); i++) {
 					if (this->towns.at(i).getPositionX() == (mouse_x / 32 * 32) && this->towns.at(i).getPositionY() == (mouse_y / 32 * 32)) {
 						this->townController = i;
 						std::cout << "\ntownCon " << this->townController;//debug
@@ -56,7 +56,7 @@ void Actor::takeControl(sf::Event event, Map& map, sf::RenderWindow& w, std::vec
 			//UNIT-TARGET--------
 		case sf::Keyboard::Right:
 			this->unitController++;
-			if (this->unitController >= this->units.size())
+			if (this->unitController >= static_cast<int>(this->units.size()))
 				this->unitController = 0;
 			break;
 			//CREATE-TOWN--------
@@ -233,10 +233,10 @@ void Actor::setTotalProdaction(int totalProdaction)
 
 void Actor::endOfTurn(Map& map)
 {
-	for (int i = 0; i < towns.size(); i++) {
+	for (int i = 0; i < static_cast<int>(towns.size()); i++) {
 		towns[i].endOfTurn(map);
 	}
-	for (int i = 0; i < units.size(); i++) {
+	for (int i = 0; i < static_cast<int>(units.size()); i++) {
 		units.at(i).recharge();
 	}
 }
