+++ sfml template/Main.cpp
@@ -2,7 +2,7 @@
 #define SUCCESS_EXIT 1
 
 int main(void) {
-	std::srand(time(NULL));
+	std::srand(static_cast<unsigned int>(time(NULL)));
 	WindowManager::mainWindow();
 	return SUCCESS_EXIT;
 }
